@@ -22,6 +22,7 @@ var processors = []processor{
 	patchProfile,
 	patchDns,
 	patchProviders,
+	patchTun,
 	validConfig,
 }
 
@@ -63,9 +64,6 @@ func patchDns(cfg *config.RawConfig, _ string) error {
 			EnhancedMode:      C.DNSFakeIP,
 			FakeIPRange:       defaultFakeIPRange,
 			FakeIPFilter:      defaultFakeIPFilter,
-			FallbackFilter: config.RawFallbackFilter{
-				GeoIP: false,
-			},
 		}
 
 		cfg.ClashForAndroid.AppendSystemDNS = true
@@ -79,7 +77,7 @@ func patchDns(cfg *config.RawConfig, _ string) error {
 }
 
 func patchProviders(cfg *config.RawConfig, profileDir string) error {
-	forEachProviders(cfg, func(index int, total int, key string, provider map[string]interface{}) {
+	forEachProviders(cfg, func(index int, total int, key string, provider map[string]any) {
 		if path, ok := provider["path"].(string); ok {
 			provider["path"] = profileDir + "/providers/" + common.ResolveAsRoot(path)
 		}
