@@ -1,3 +1,5 @@
+//go:build premium
+
 package main
 
 //#include "bridge.h"
@@ -22,8 +24,7 @@ func init() {
 		sub := log.Subscribe()
 		defer log.UnSubscribe(sub)
 
-		for item := range sub {
-			msg := item.(*log.Event)
+		for msg := range sub {
 
 			cPayload := C.CString(msg.Payload)
 
@@ -49,12 +50,7 @@ func subscribeLogcat(remote unsafe.Pointer) {
 		sub := log.Subscribe()
 		defer log.UnSubscribe(sub)
 
-		for i := range sub {
-			msg, ok := i.(*log.Event)
-			if !ok {
-				continue
-			}
-
+		for msg := range sub {
 			if msg.LogLevel < log.Level() && !strings.HasPrefix(msg.Payload, "[APP]") {
 				continue
 			}
